structure created

+++ DTOs/CartItemDTO.cs
@@ -3,6 +3,6 @@ namespace Loop.DTOs.Common
     public class CartItemDTO
     {
         public int Quantity { get; set; } = 0;
-        public int Price { get; set; } = 0;
+        public float Price { get; set; } = 0;
     }
 }
