+++ DTOs/OrderItemDTO.cs
@@ -3,7 +3,7 @@ namespace Loop.DTOs.Common
     public class OrderItemDTO
     {
         public int Quantity { get; set; } = 0;
-        public int Price { get; set; } = 0;
+        public float Price { get; set; } = 0;
         public int PaymentStatus { get; set; } = 0;
     }
 }
