+++ DTOs/PaymentDTO.cs
@@ -4,7 +4,7 @@ namespace Loop.DTOs.Common
     {
         public string Provider { get; set; } = "";
         public string ProviderRef { get; set; } = "";
-        public int Amount { get; set; } = 0;
+        public float Amount { get; set; } = 0;
         public int Status { get; set; } = 0;
         public DateTime PaidAt { get; set; }
     }
