+++ DTOs/PayoutDTO.cs
@@ -4,7 +4,7 @@ namespace Loop.DTOs.Common
     {
         public DateTime From { get; set; }
         public DateTime To { get; set; }
-        public int Amount { get; set; } = 0;
+        public float Amount { get; set; } = 0;
         public int Status { get; set; } = 0;
     }
 }
