+++ Models/Order.cs
@@ -8,7 +8,7 @@ namespace Loop.Models.Common
         public Guid Id { get; set; } = Guid.NewGuid();
         public Guid CustomerId { get; set; }
         public Guid AddressId { get; set; }
-        public int TotalAmount { get; set; } = 0;
+        public float TotalAmount { get; set; } = 0;
         public int Status { get; set; } = 0;
         public int PaymentStatus { get; set; } = 0;
         public int ShippingStatus { get; set; } = 0;
